# server/docs/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "SCIM API documentation for Gluu. For more details please visit https://tools.ietf.org/html/draft-ietf-scim-api-19",
    "version" : "v2",
    "title" : "Gluu SCIM 2.0 Apis Documentation",
    "termsOfService" : "https://github.com/GluuFederation/",
    "contact" : {
      "name" : "Rahat Ali",
      "url" : "http://gluu.org",
      "email" : "rahat@gluu.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "gluu.org",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/scim/v2/Bulk" : {
      "post" : {
        "tags" : [ "v2Bulk" ],
        "summary" : "Bulk Operation",
        "description" : "SCIM Bulk Operation (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.7)",
        "operationId" : "bulkOperation",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "BulkRequest",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/BulkRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BulkResponse"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      }
    },
    "/scim/v2/Groups" : {
      "get" : {
        "tags" : [ "v2Groups" ],
        "summary" : "List Users",
        "description" : "Returns a list of Groups (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.4)",
        "operationId" : "listGroups",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "filter",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortOrder",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ListResponse"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Create Group",
        "description" : "Create Group (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "createGroup",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Group",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Group"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      }
    },
    "/scim/v2/Groups/{id}" : {
      "get" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Find Group by id",
        "description" : "Returns a Group on the basis of provided id as path param (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.4.2.1)",
        "operationId" : "getGroupById",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Update Group",
        "description" : "Update Group (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "updateGroup",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Group",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Group"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Delete Group",
        "description" : "Delete Group (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "deleteGroup",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      }
    },
    "/scim/v2/Users" : {
      "get" : {
        "tags" : [ "v2Users" ],
        "summary" : "List Users",
        "description" : "Returns a list of users (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.4)",
        "operationId" : "listUsers",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "filter",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortOrder",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ListResponse"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "v2Users" ],
        "summary" : "Create User",
        "description" : "Create User (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "createUser",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "User",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      }
    },
    "/scim/v2/Users/{uid}" : {
      "get" : {
        "tags" : [ "v2Users" ],
        "summary" : "Find User by id",
        "description" : "Returns a Users on the basis of provided uid as path param (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.4.2.1)",
        "operationId" : "getUserByUid",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "v2Users" ],
        "summary" : "Update User",
        "description" : "Update User (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "updateUser",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "User",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v2Users" ],
        "summary" : "Delete User",
        "description" : "Delete User (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "deleteUser",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      }
    }
  },
  "definitions" : {
    "Address" : {
      "properties" : {
        "operation" : {
          "type" : "string"
        },
        "primary" : {
          "type" : "boolean",
          "default" : false
        },
        "formatted" : {
          "type" : "string"
        },
        "streetAddress" : {
          "type" : "string"
        },
        "locality" : {
          "type" : "string"
        },
        "region" : {
          "type" : "string"
        },
        "postalCode" : {
          "type" : "string"
        },
        "country" : {
          "type" : "string"
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
    "BulkOperation" : {
      "properties" : {
        "bulkId" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "method" : {
          "type" : "string"
        },
        "path" : {
          "type" : "string"
        },
        "location" : {
          "type" : "string"
        },
        "data" : {
          "type" : "object"
        },
        "status" : {
          "type" : "string"
        },
        "response" : {
          "type" : "object"
        }
      },
      "xml" : {
        "name" : "BulkOperation"
      }
    },
    "BulkRequest" : {
      "properties" : {
        "schemas" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "failOnErrors" : {
          "type" : "integer",
          "format" : "int32"
        },
        "operations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/BulkOperation"
          }
        }
      }
    },
    "BulkResponse" : {
      "properties" : {
        "schemas" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "operations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/BulkOperation"
          }
        }
      }
    },
...
    "Email" : {
      "properties" : {
        "operation" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "display" : {
          "type" : "string"
        },
        "primary" : {
          "type" : "boolean",
          "default" : false
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
...
    "Group" : {
      "required" : [ "schemas" ],
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "externalId" : {
          "type" : "string"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        },
        "schemas" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "displayName" : {
          "type" : "string"
        },
        "members" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/MemberRef"
          }
        }
      }
    },
    "GroupRef" : {
      "properties" : {
        "value" : {
          "type" : "string"
        },
        "display" : {
          "type" : "string"
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
    "Im" : {
      "properties" : {
        "operation" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "display" : {
          "type" : "string"
        },
        "primary" : {
          "type" : "boolean",
          "default" : false
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
    "ListResponse" : {
      "properties" : {
        "totalResults" : {
          "type" : "integer",
          "format" : "int32"
        },
        "startIndex" : {
          "type" : "integer",
          "format" : "int32"
        },
        "itemsPerPage" : {
          "type" : "integer",
          "format" : "int32"
        },
        "schemas" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "resources" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Resource"
          }
        }
      },
      "xml" : {
        "name" : "Resources"
      }
    },
    "MemberRef" : {
      "properties" : {
        "operation" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "display" : {
          "type" : "string"
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
    "Meta" : {
      "properties" : {
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModified" : {
          "type" : "string",
          "format" : "date-time"
        },
        "location" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "attributes" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "resourceType" : {
          "type" : "string"
        }
      }
    },
    "Name" : {
      "properties" : {
        "formatted" : {
          "type" : "string"
        },
        "familyName" : {
          "type" : "string"
        },
        "givenName" : {
          "type" : "string"
        },
        "middleName" : {
          "type" : "string"
        },
        "honorificPrefix" : {
          "type" : "string"
        },
        "honorificSuffix" : {
          "type" : "string"
        }
      }
    },
    "PhoneNumber" : {
      "properties" : {
        "operation" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "display" : {
          "type" : "string"
        },
        "primary" : {
          "type" : "boolean",
          "default" : false
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
    "Photo" : {
      "properties" : {
        "operation" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "display" : {
          "type" : "string"
        },
        "primary" : {
          "type" : "boolean",
          "default" : false
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
    "Resource" : {
      "required" : [ "schemas" ],
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "externalId" : {
          "type" : "string"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        },
        "schemas" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "User" : {
      "required" : [ "schemas" ],
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "externalId" : {
          "type" : "string"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        },
        "schemas" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "userName" : {
          "type" : "string"
        },
        "name" : {
          "$ref" : "#/definitions/Name"
        },
        "displayName" : {
          "type" : "string"
        },
        "nickName" : {
          "type" : "string"
        },
        "profileUrl" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "userType" : {
          "type" : "string"
        },
        "preferredLanguage" : {
          "type" : "string"
        },
        "locale" : {
          "type" : "string"
        },
        "timezone" : {
          "type" : "string"
        },
        "active" : {
          "type" : "boolean",
          "default" : false
        },
        "password" : {
          "type" : "string"
        },
        "emails" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Email"
          }
        },
        "phoneNumbers" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PhoneNumber"
          }
        },
        "ims" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Im"
          }
        },
        "photos" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Photo"
          }
        },
        "addresses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Address"
          }
        },
        "groups" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/GroupRef"
          }
        },
        "entitlements" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Entitlement"
          }
        },
        "roles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Role"
          }
        },
        "x509Certificates" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/X509Certificate"
          }
        },
        "customAttributes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CustomAttributes"
          }
        }
      },
      "xml" : {
        "name" : "User"
      }
    },
...
  }
}